error_analysis/requirements.txt
# Error Analysis and Interpretability Framework Requirements

# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0

# Image processing
Pillow>=8.3.0
scikit-image>=0.18.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Progress tracking
tqdm>=4.62.0

# Logging and utilities
pathlib2>=2.3.0

# Optional: For advanced clustering
# umap-learn>=0.5.0
# hdbscan>=0.8.0
